+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New method `C::getAllAppModules()` to only get app-related loaded modules like the App module, no kernel modules.
 - Slugifying is improved and can now create safe filenames via `C::Formatter()->slugify($text, true)`.
 - Simple method to download files via `C::Http()->download($url)`.
+- The model cache can easily be cleared via `$model->clearModelCache()` or to clear the full cache of all entities
+  of this model, use `ModelClass::clearModelsCache()`.
 
 ### 🔧 Changed